+++ src/os2datascanner/engine2/model/derived/__init__.py
@@ -1,6 +1,7 @@
-from . import pdf # noqa
-from . import tar # noqa
-from . import zip # noqa
-from . import mail # noqa
-from . import filtered # noqa
-from . import libreoffice # noqa
+from . import pdf  # noqa
+from . import tar  # noqa
+from . import zip  # noqa
+from . import mail  # noqa
+from . import filtered  # noqa
+from . import libreoffice  # noqa
+from . import spreadsheet  # noqa